requirements.txt
numpy
pillow

icalendar